Reorganize code structure and update CMakeLists.txt. Ubuntu compile fix.
