+++ src/config/config.h
@@ -22,7 +22,7 @@
 #include <cpptoml.h>
 #include <pwd.h>
 #include <spdlog/spdlog.h>
-#include <filesystem>
+#include "utils/filesystem.h"
 #include <iostream>
 #include <map>
 #include <memory>
