+++ src/daemon/daemonizer.cpp
@@ -16,7 +16,7 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include "daemonizer.h"
+#include "daemon/daemonizer.h"
 
 /**
  * Forking logic for classic style daemon functionality
