+++ src/utils/string-from-char.cpp
@@ -16,7 +16,7 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include "util.h"
+#include "utils/string-from-char.h"
 
 /**
  * @brief Safely converts a char array to a std::string
